Create modelA.ipynb, rework stmt_emb.py
modelA.ipynb is the data loading and EDA portion
of the model. There's still a long way to go to
get an operation model.

statement_embedding.py now searches for labels
(and their inverses) using a dict, just like it
already did for statements. Perhaps it's not ideal
to create the dict everytime the function is
called